requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-jose
sqlalchemy
asyncpg
alembic
passlib
pydantic
python-dotenv
ydantic-settings